Fixed all current bugs
* Added support for the waza_p2.bin files in the pokemon XML data.
* Added XML attribute for indicating the game version inside pokemon xml
files, given those are trickier to handle.
* Fixed issues with the waza_p.bin and waza_p2.bin files either missing
entries or having too many. Those were related to the difficulty of
determining whether the data was for EoS or EoT/D.
* Abandoned replacing move ids with actual move names, as several moves
share the same move names!
* The base form index, and the pre-evo index are still very unintuitive,
as both of these for the secondary gender are in the 600+.. But the
original index of each entities isn't stored in memory, so its hard to
do something better with this right now..