Added new feature to audioutil
* Added a way to write a "blank" populated midi conversion info file for
a set of loaded swd files. Essentially, it will look at all the
presets/programs in all the loaded the swd files, and build a XML file
following the structure of the conversion info file. So that means end
users have a lot less XML to write, and can just edit the values for a
given DSE track, and load that cvinfo at the command line to get their
smd files converted to midi by the parameters they want.
* Added a basic statistic collector for the DSE batch loader. Hopefully,
I'll put everything in a definition file eventually..
- I found out that, accurate conversion to midis and soundfont won't be
ever a thing. After managing to bake the samples correctly, I noticed
that, the release param played a way bigger role than expected. Then,
the LFOs in the soundfont format can only oscillate up to 100hz, while
the LFO in the DSE format can go up to much higher frequencies. Baking
LFO effects is out of the question because resampling will modify the
perceived LFO oscillation rate and just sound weird. I could render
every single notes with the correct frequency and etc, but that would
make the soundfont enormous, possibly beyond the maximum size.. So, I'll
probably have to scrap that part, we'll see.. I'm also beginning work on
a real-time player for DSE music tracks, so that code I made might be
useful.