SF2 handling modified the way loop legality fix is handled
* Made it so the samples are looped fully, using their original loop
points, several times over until the sample is able to be looped by the
SF2 format.
The old method was basically copying tiny bits of samples around the
loop points, and it was just dumb.

Ideally, this should be handled by whatever put the samples in the SF2.
But for now, this will have to do.

- The fix isn't without issues. Some samples now have audio artifacts.
But I believe it might come from another component in audioutil. The old
way the loop legality was fixed might have covered up the issue for all
this time.