Added support for bgm and sd container for DSE sequenced music
* Added support for loading and exporting any "bgm container" at the
command line. Basically, those are SIR0 wrapped pair of a smdl and swdl.
It also can find out if a bgm container has a sedl instead of a smdl,
and smartly avoid parsing it. The user must specify the file extension
of the bgm container files to be expected.

* Changed the way optional commandline parameters are parsed. They now
require a function that returns whether, according to the current option
and parameters parsed, the parameter is optional or required.
* Updated Audioutil, GFXCrunch, and StatsUtil to use the new system, and
hopefully fixed most bugs with it.
* GFXCrunch and statsutil will receive a major update to their command
line param handling soon too. It got too complex for nothing. And
Statsutil doesn't need to figure out a locale as much as it used to! It
was a dumb assumption on my part.