Blob, SADL support added. Fixed part of content analyzer dependency.
* Added the macro "USE_PPMDU_CONTENT_TYPE_ANALYSER" to enable/disable
the content_type_analyzer.hpp dependency that a lot of files have.
Should make it a bit easier to re-use the code in something else without
having to carry all that junk around. Also tweaked all projects using
the content type analyzer to declare that macro properly.
* Centralized the magic number constants for the main DSE containers
smdl, swdl, sadl, and sedl. This way we avoid redefinitions and circular
includes.
* Specialized the bytes to integer functions to eliminate the annoying
if block that could be avoided at runtime. (Still need to improve on
that stuff though)
* Added SADL header. Will need to eventually implement parsing. Though,
its not a priority in this project. I might move the DSE stuff to its
own project.
* Added "blob" loading support for DSE audio. Still pretty buggy.
Hopefully it won't break anything..
* Changed lowest attenuation for DSE music to 25 decibels from 19. Seems
to somewhat improve things a bit this far. Will eventually need to
change this system to use something more accurate to convert
proportional volume to decibel attenuation. But lack of data on the
DSP's max output loudness makes it a bit hard..
* Various misc fixes.