Added Multi-DSE Version Support
* Tweaked the code so that most of the needed information can be parsed
from DSE v0x402 files as well as the more common v0x415 version. Its
very rudimentary and ugly. But, I can't figure out a way to do anything
smart in this case, short of rewriting everything..
* Fixed a lot of objects' "Read" parsing methods so that they check the
iterator range they're in to avoid going out of bounds and doing some
nasty stuff.. Sometimes validating before crashing is just not worth it,
so at least crashing gracefully if the file is corrupted is a better
option ^^;
* Added support for exporting samples without converting them to pcm16.
So adpcm and pcm8 files are exported as is. pcm8 are exported as pcm8
wave files, while adpcm are exported as-is with their preamble.
* Temporarily fixed the SWDL, XML export to work for generic program and
split. I doubt we'll ever really need to export/import any version
specific parameters, but maybe set something in place just in case?
* Moved all DSE version-specific structs and class to the smdl.cpp and
swdl.cpp files to keep them from cluttering the headers.
- I need to simplify the byte to integer reading functions I made and
place them in their own header, to make it easier to re-use the code
without including all that messy "utils" lib I made.
- I think I figured a way to preserve loop accuracy better. And it would
involve processing samples from their original format into pcm16. Since
ADPCM loop points are counted in blocks of 4 bytes, this means we have
theoretically way less drift in the position of the loop points than if
we'd just convert to pcm16 and convert the loop point position! But I
still need to implement that, I'll need to fine-tune a "real-time" adpcm
converter class for that to work.