Audioutil - Added Verbose + Event 0xBE is Set Mod
* Modified the DSE event interpreter so that skip byte events are
properly ignored now, and don't count as unknown events. We don't need
to do anything with them, since the bytes to skip are grabbed as
paramters when parsing the file. So nothing to do here!
* Event 0xBE is now Set Modulation. Should have done that long ago, but
I kept forgetting about it.
* Added a check in the DSE event interpreter so that note number higher
than 11 get logged and are properly ignored.
* Moved most of the log output to verbose. Since most of it is
superfluous.
* Added "-v" commandline switch to toggle verbose log output on.
* Made Event 0xCB SkipNext2Bytes1.
* Tweaked a lot of comments in random places.
* Modified DSE BlobScanner so it lists the file offsets where the DSE
containers were found, on screen and in the log file.