Statsutil : Added basic ssb support
Statsutil mostly compiles, but probably doesn't works properly.
* Added WIP ssb support
* Added WIP basic script loading
* Added WIP Mappa_*.bin file handing and a storage class for the parsed
data.
* Added WIP base for editing game overlays and arm9 binary.
* Added proper game and local detection that's not tied to the GameStat
class! (Need to actually remove that  from the gamestat class, because
its redundant)
- Caused the compiler to HATE statsutil. Now it crashes while compiling
it with optimizations on. Its a bug in the VS2015 update 2 compiler. I'm
seriously considering ditching the MSVC compiler and go with mingw or
clang.. The project and directory structure has gotten messy and ugly
anyways.