Fix for issue  #26 with SIR0 content being incorrectly padded
- Fixed issues #26 where the stat growth was nulled out for most pokemon
past a certain point. It was due to the length of the padding being
calculated using "data.size() % 16", which gives the remainder and not
the padding required to append. In order to fix the issue, I simply made
a function that return an aligned value on demand which handles both
cases where the length is aligned, and cases when the value is not
aligned, correctly.
- Made a few exceptions strings for some of the error cases clearer.
- Had to disable optimisation for the statsutil release build. Because
there's a very hard to  trace compiler bug causing the compiler to crash
while compiling that project in release.