Added support for a few more parameters + various fixes
* Fixed NULL level id handling.
* Fixed 14bits integer handling.
* Added a base class for 14bits integer. Not yet tested or in use at the
moment.
* Added some nested throws when doing a batch script export or import!
This way, the file being processed is acknowledged in the error message.
* Added automated detection of single scripts from drag and drop. Either
raw ssbs or xml. XML will detect the game version and region, but the
ssb is EXPLORERS OF SKY ONLY for now, which should get fixed somehow
eventually?
* Added WIP XML template for script related data, so that, the data can
be loaded from there instead of in a big constant array inside the code!