statsutil: Introduced typed routine, reworked and simplified some of the decompiler + parser
* Made it so the type of a routine, except for aliases, is stored in the
node name. It reduces a bit the clutter, and might make things faster.
* Removed the useless parameter value for routines types for which it
was always 0.
* Also added the command line parameter "scriptdebug" to enable by
default all debug branch commands. It simply inverts their condition. So
debug branch command that in a retail build jump over debug code, are
then not jumping in the  retail build, but in the debug build. Should
make hunting for debug scenes a bit faster!
* Renamed and cleaned up a lot of things
* Added basic handling for menuids, station and acting layer ids. Still
not sure what they refer to yet..