Statsutil: Fixed error reporting on script compilation.
* Added output of compiler report to the compiler_report.txt file,
outputed in the current working directory. Options to change this will
be coming eventually if needed!
* Added a way for the parser to report warnings about syntax of the
scripts and etc.
* Added better organized output of all error messages.
* Show on what line the error is, along with the line in question.
* Changed the default setting for pugixml to indent using spaces instead
of tabs. In order to make the error reporting more accurate!(Tabs don't
have a fixed size, even though the character for it is still one byte!)
* Modified the async task handler so it actually push exceptions on a
queue instead of just stopping silently, and sending the exception to
exception hell.